build.have_clang() no longer deadlocks on call; Fixes #8
`have_clang()` uses subprocess.call() and pipes stdin.  But call()
creates a Popen object and calls wait() on it.  The python docs specify
that wait() can deadlock if there is a lot of output piped to stdout without
anything reading from it.  By sending the output to /dev/null, the deadlock
will not occur.